# More efficient encoding of lists