Tested with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                      (size)   Mode  Cnt         Score         Error  Units
ListOfBooleansWriting.zioJson       1  thrpt    5  42626498.232 ± 1191762.765  ops/s
ListOfBooleansWriting.zioJson       3  thrpt    5  30622270.257 ±  565535.698  ops/s
ListOfBooleansWriting.zioJson      10  thrpt    5  11410538.676 ±   52518.404  ops/s
ListOfBooleansWriting.zioJson      33  thrpt    5   3954503.595 ±   10465.811  ops/s
ListOfBooleansWriting.zioJson     100  thrpt    5   1141918.106 ±   78351.131  ops/s
ListOfBooleansWriting.zioJson     333  thrpt    5    356808.175 ±    8228.529  ops/s
ListOfBooleansWriting.zioJson    1000  thrpt    5    105240.520 ±     537.849  ops/s
```

#### After
```txt
Benchmark                      (size)   Mode  Cnt         Score        Error  Units
ListOfBooleansWriting.zioJson       1  thrpt    5  42893766.949 ± 830789.797  ops/s
ListOfBooleansWriting.zioJson       3  thrpt    5  31424740.861 ± 205664.394  ops/s
ListOfBooleansWriting.zioJson      10  thrpt    5  14039622.426 ± 292832.648  ops/s
ListOfBooleansWriting.zioJson      33  thrpt    5   4321492.200 ±  74325.392  ops/s
ListOfBooleansWriting.zioJson     100  thrpt    5   1406430.432 ±  51164.812  ops/s
ListOfBooleansWriting.zioJson     333  thrpt    5    407791.373 ±   6986.114  ops/s
ListOfBooleansWriting.zioJson    1000  thrpt    5    123098.798 ±   5591.825  ops/s
```